Change AppLauncher.launch to take an app class rather than an instance.
The old AppLauncher.launch API would take in input an Application instance,
rather than an Application class.
This can be confusing, because this way the instance will be created before
the parsing of @CmdLine Args.

This change makes so that the all AppLauncher.launch methods take in input an
Application class, and AppLauncher.launch creates an instance of this class
after having parsed the @CmdLine Args.
All code that was relying on the old signature has also been modified.