add daemon and ephemeral processes to thermos adds the following fields to the thermos Process schema:
+  daemon        = Default(Integer, 0) # boolean
+  ephemeral     = Default(Integer, 0) # boolean
+  min_duration  = Default(Integer, 5) # integer seconds

daemon processes are processes that will be rescheduled even if they
succeed.  they are still bound by max_failures, so the user must set
max_failures=0 if they want it to run forever.

min_duration is the minimum time between runs to wait.  set to 5 by default,
maybe should be much higher...should probably put in a hardcoded lower
limit.

ephemeral processes are processes that are not taken into consideration when
determining whether or not the thermos task is complete.  for example,
"background" processes such as checkpoints, log rotates, log saves and the
like should be ephemeral (and/or ephemeral+daemon) processes.

factors out some of the more specific task planning logic into the
TaskPlanner.  this actually simplifies the runner a bit, and makes it a
little more clear how to split out the process management stuff into
something more testable (i.e.  non-forking.)