Move the Thermos GC Executor to being an online process
Previously, every time the scheduler received an offer from a slave, it would attempt to launch a GC task on that slave, if a timeout guarantee is met (i.e. a GC task hadn't been launched for X minutes). Since the scheduler created a unique executorID for every GC task invocation, the slave would create a new executor every time. This change removes the executorID uniqueness; thus, if a slave is already running a GC executor, that executor will simply receive a new launchTask() command with the new set of retainedTasks. It also updates the GC executor to be long-lived and exist beyond the lifetime of a single launchTask.

The current implementation uses a queue inside the GC executor to handle multiple concurrent launchTasks, i.e. if the executor receives a launchTask before the previous run has completed. This queue is limited to length 1: if a GC executor receives a task while it already has one queued, the queued task is discarded and a FINISHED message sent to the scheduler. In this way, only the most recent AdjustRetainedTasks message is ever used.