upgrade to pystachio 0.7.1
i killed the @Provided annotations in pystachio 0.7.0 because they add a ton
of complexity for little gain.  killing them will make it easier to iterate
more quickly on pattern matching and config plugins which are part of the
yak shave to make it simpler to support pulling notification configs into
.mesos files for d2a.

this change is three parts:
  * kill mesos.package and related deprecation warnings -- there were only a couple configs that specified mesos.package and afaict all of them went unused
  * rewrite {{thermos.user}} => {{role}}, {{mesos.role}} => {{role}}, {{mesos.cluster}} => {{cluster}}
  * filter {{thermos.task_id}} and {{thermos.ports}} as acceptable unbound references inbound to the thermos runner

so far as i can tell this change is fully compatible with the executors
running in production, and executors built against this release *should*
work with outdated clients.