dynamically load schemas
Until we have a common aurora config tree e.g. src/aurora, we'll need to
split schemas into OSS and internal counterparts.  This does that.  Please
advise on the approach.  It's sort of ugly (see the introspection in
AuroraConfigLoader to see what I mean) but I can't think of anything better
that satisfies the short term need of opensourcing the client.