Building tests for aurora client commands, round 1.
This is a proof of concept of the mock approach to building client command-line
tests. It builds up a mock API, and ensures that the client performs the
correct API calls, and then does the correct validity tests on the results.

This only implements one test, of a successful job creation. If this approach
looks good, I'll go ahead and fill in a ton more similar tests - of other
cases of create, and similar structures for other commands.