Added getJobSummary call to thrift API
Added a new Thrift service to host all read only api's. Added a getJobSummary
call to the read only service end point. The output of the call will be used to
render the home page.

Implemented the getJobSummary API in SchedulerThriftInterface and added a
corresponding unit test. To keep it simple, the implementation of getJobSummary
is a clone of SchedulerzHome.get().

Testing Done:
gradle clean build

Reviewed at https://reviews.apache.org/r/16231/