Ensure SchedulerActive event is dispatched before leader is advertised.
I dug into this after noticing SchedulerIT test flakiness (deadlock) as a result
of tearDown actions being executed SchedulerActive event was being dispatched.
The test is arranged such that it expects the effects of that event to occur
before the test exits.  The deadlock was related to EasyMock's default global
lock for mock answering, so that particular deadlock does not affect us in
production.

Reviewed at https://reviews.apache.org/r/16636/