Begin cleanup of StateManager interface by removing state change via Query.
State changes via a broad query is a rarely-used feature, and probably usually
not a good idea.  The query-based API was useful because it sometimes saves
code at the call site, but the big downside is that it makes it more difficult
for StateManager to perform course-correcting operations.
The test case added to StateManagerImplTest illustrates a now-fixed regression
that could contribute to state drift between the scheduler and the outside
world.

Bugs closed: AURORA-27

Reviewed at https://reviews.apache.org/r/16767/