Add support for global hooks to the client.
To support site-specific policy requirements, we'd like to be able to
have hooks that are applied to every configuration, and that are used
regardless of whether a configuration specifically enables hooks. We'd
like these to be baked in to the client.

This change adds a global registry, which can be initialized with a
set of global hooks. In case of errors/emergencies, there is an
override flag to disable the global hooks.

Bugs closed: aurora-269

Reviewed at https://reviews.apache.org/r/19411/