Implement client command logging with a unique token.
Adding a centralized logger interface for clientv2.

For the aurora analytics work, we need some hooks for logging. We'd
like all client logging to go through a common interface, and for
client logs that might be sent to the centralized logging service to
be tagged with identifiers and userids.

This change implements a simple minimal logging function that's used
by clientv2 logging calls, and which attaches a universally unique
client identifier to each log message.

Bugs closed: aurora-276

Reviewed at https://reviews.apache.org/r/19396/