Modify the way that config binding helpers get registered.
Config binding helpers (components that add macros to the pystachio
config language) self-registered in a way that made it difficult to
provide parameters to initialize them.

This change switches to an explicit construction/registration, instead
of auto-construction when the class is registered. (Interestingly, this
is the way that the documentation on the binding helpers code says that
it works!)

With this change, instead of writing:

     FooHelper.register()

You write:
    BindingHelper.register(FooHelper())

Which makes it possible to do:
   BindingHelper.register(FooHelper(url=bar))

Testing Done:
All client unit tests run and passed.

Bugs closed: aurora-496

Reviewed at https://reviews.apache.org/r/22082/