Implement a TRequestsClient as a prelude to kerberization.
This is the first bit of AURORA-515.  To add kerberos support, we just need
to add a dependency on requests_kerberos and inject KerberosAuth() as the
'auth=' parameter to TRequestsClient with the proper service principal
specified.

Testing Done:
Updated client/api tests for the new transport.

Reviewed at https://reviews.apache.org/r/22280/