Make testing of v1 command hooks easier.
In order to be able to test v1 command hooks on real commands, it would
be helpful to be able to mock out the body of the commands. In order to
do this, I've separated the commands into an invocation frame, and
the actual command implementation.

Bugs closed: aurora-603

Reviewed at https://reviews.apache.org/r/23949/