Fix bug in deprecation warning for cancel-update.
While working on improving the coverage of our end-to-end tests,
I discovered a cut-and-paste error in the implementation of the
warning message for cancel_update.

This is a quick-fix for this problem. Testing for this will be included
in the testing coverage changes that I'm working on.

Bugs closed: aurora-650

Reviewed at https://reviews.apache.org/r/24651/