Implement server-driven update commands.
This change contains the basic commands needed to work with the
scheduler-driven updater. (It does not yet cover querying for the status
of the update; that will come in a subsequent change.)

Testing Done:
New suite of tests for the new command; all unit tests pass.

Reviewed at https://reviews.apache.org/r/25255/