Make thermos more cognizant of user deletions
This is changes 1 of 2 for AURORA-175.  The second change adds an exit
status contract between Aurora and Thermos.

This change allows a process in the Thermos state machine to go directly
from WAITING -> FAILED if the user associated with that process has been
deleted.  It also persists a 'uid' field to the Thermos RunnerHeader so that
we can check against process UIDs which in theory have higher fidelity than
usernames.

Bugs closed: AURORA-175

Reviewed at https://reviews.apache.org/r/25972/