Don't kill GC Executor after period of inactivity
The GC executor is configured to exit after 15 minutes of
inactivity. This leads to a race where the mesos slave gets a
launchTask message for a GC executor just as the executor has exited,
causing TASK_LOST noise. This also increases the risk that a slave
will lose its GC executor due to the scheduler not being able to find
a slot for it (since GC executors will have a higher churn rate).

Cluster operators will still be able to deploy new versions of the
GC executor as the 24-hour max lifetime limit is still in place. This
patch only removes the inactivity limit.

Testing Done:
./pants src/test/python/apache/aurora/executor:gc_executor

Bugs closed: AURORA-788

Reviewed at https://reviews.apache.org/r/26300/