Use virtualenv to build pants instead of pex.
The goal of this change is to reduce CI flakiness by using `pip
install` rather than `pex` to bootstrap pants. This still reaches out
to external servers (and thus has the potential to be flaky) but is
at least configurable (via `--default-timeout`). In a future review
we can consider mirroring to svn.apache.org and disabling PyPI lookups.

This requires a change to production executor code to use
`sys.executable` rather than chmod+x-ing the runner PEX. Failure to
do this results in the pants virtualenv's `site-packages` shadowing
the dependencies in the `pex` (fairly obviously as there is a test
failure related to psutil1/2 conflicts). I'm sure I'm papering over
a PEX bug here but IMO this is fine as the only reason for it is to
execute the thermos runner PEX under a different python interpreter
than the executor is using (based on the hash-bang in the pex). This
is not a configuration we need to support IMO.

Testing Done:
git clean -fdx
./pants src/test/python:all

Bugs closed: AURORA-876

Reviewed at https://reviews.apache.org/r/27009/