get_client_version should never be allowed to run unpatched in test.
get_client_version tries to get the PEX-INFO of sys.argv[0], which is
usually pants when running tests (!) if the format of PEX-INFO changes to a
new version of pants (as is the case for pants 0.0.25), then tests will
start failing with an inscrutible error.  instead, just patch
get_client_version to something sensible.

Reviewed at https://reviews.apache.org/r/29132/