Ensure TaskStateChange event emitted by StateManagerImpl has valid data.
The `StateManagerImpl` class emits a pubsub event when a task transitions to the
ASSIGNED state but before it attaches the associated slave id and host. This
patch ensures that the pubsub event has the associated slave id and host when
emitted.

Testing Done:
./gradlew build -Pq

Bugs closed: AURORA-1016

Reviewed at https://reviews.apache.org/r/29915/