Revert backwards incomaptible chnages to GC ExecutorInfo.
The Docker patch (7ba6226) started to set the shell attribute of of
ExecutorInfo. This breaks backwards compatability for the GC executor as it will
continue to have the same id but different info now. In addition it is not
necessary to set this value because it defaults to true.

Testing Done:
./gradlew build -Pq

Bugs closed: AURORA-1076

Reviewed at https://reviews.apache.org/r/30471/