Removing redundant scheduling loop in preemptor.
Context is that the removed code was attempting to mitigate
a race where we preempt a task unnecessarily because new offers
have arrived.  Given that preemption begins immediately after
we have passed over the offers, this additional pass is redundant.
Additionally, the race still exists (an offer could arrive
right after we preempt), so the mitigation is of dubious value.

Bugs closed: AURORA-1156

Reviewed at https://reviews.apache.org/r/31508/