Remove url related methods out of AuroraCommandContext
This is a refactor of AuroraCommandContext which removes all url related methods
out of it. The objective of this refactor is to remove functionality from
AuroraCommandContext to allow for easier testing of commands. This commit also
adds two tests for commands which were using the url related functionality but
lacked test coverage.

Testing Done:
./pants test.pytest --no-fast src/test/python/apache/aurora/client/cli::

Reviewed at https://reviews.apache.org/r/32900/