Add typed Shiro permissions SPI.
Working on an implementation of a Shiro Realm to replace the old
CapabilityValidator I realized we're missing some information with
the new API. This patch allows a Realm implementation to optionally
introspect a permission check for Aurora-specific information about
exactly what's being attempted, while maintaining compatibility
with Shiro realms that don't know anything about Aurora, such as the
included IniRealm.

To do this I've added a new SPI package, and documented
backwards-compatibility considerations with it. The idea is that
a third party can write an Aurora-aware Shiro Realm module against
a stable ABI and gets one release to update it if we want to make
changes to it.

Testing Done:
./gradlew -Pq build

Reviewed at https://reviews.apache.org/r/33411/