Removes version on virtual pacakge in debian control
java-runtime-headless is a virtual package and cannot have a version
specified. The consequence of this is that the built packages require openjdk-8
be installed which prevented the packages from working with oracle-java8.
java8-runtime-headless is a virtual package provided by both openjdk-8
and oracle-java8 so the resulting debian packages can be installed on a
machine with either versions of java8 installed.

Testing Done:
Started two docker containers, one with oracle-java8 and the other with
openjdk-8. I was able to install the aurora-scheduler package in both containers
with this change.

Bugs closed: AURORA-1400

Reviewed at https://reviews.apache.org/r/36511/