Make Storage.Work an interface instead of an abstract class.
Make `Storage.Work` and friends interfaces instead of abstract classes.

This allows you to pass lambdas and method references to calls to
`storage.work`. The main ripple is that as interface methods, `apply`
and `execute` must be public. This was done programmatically with a sed
script.

Testing Done:
./gradlew -Pq build

Reviewed at https://reviews.apache.org/r/36713/