Stop in-progress instances from pulsing when update has been aborted.
Also removed the "INSTANCE-" prefix from tooltip as it seemed superfluous and
added "ABORTED" to the tooltip for instances that were "UPDATING" when the
update was aborted. An argument can be made for adding "ABORTED" to the tooltip
for all non-terminal states (PENDING, etc.), let me know if you think we
should add it for those as well.

Reviewed at https://reviews.apache.org/r/38010/