Ignore all SessionKeys.
This change ignores all SessionKeys. Since the scheduler was doing
"requireNonNull" on SessionKey before the client needs to continue
to set SessionKey to a non-null value for this release. In the next
release the parameter can be dropped completely.

Apologies for the large diff, thankfully it's mostly red. Coverage in
SchedulerThriftInterface is very high, most of the deleted tests deal
with branches related to authentication or authorization failures,
as these are now enforced in a different layer.

Testing Done:
./graldew -Pq clean build

Reviewed at https://reviews.apache.org/r/39525/