Upgrade to pants `0.0.59` to avoid pytest errors.
The problem is described here:
  https://github.com/pantsbuild/pants/issues/2566

Changelogs are here:
  https://pypi.python.org/pypi/pantsbuild.pants/0.0.59

 pants.ini | 2 +-
 1 file changed, 1 insertion(+), 1 deletion(-)

Testing Done:
Before the fix stumbled into this:
```
$ ./build-support/jenkins/build.sh
...
11:22:51 00:00   [test]
11:22:51 00:00     [run_prep_command]
11:22:51 00:00     [test]
11:22:51 00:00     [pytest]
11:22:51 00:00       [run]

11:23:04 00:13         [chroot]
11:23:11 00:20   [complete]
               FAILURE

Exception message: Could not satisfy all requirements for pytest>=2.6,<2.7:
    pytest>=2.6,<2.7, pytest>=2.8.0(from: pytest-timeout)
```

That triggered a bit of a pants release firefight leading
to this RB and one for Medium's mono repo as well as patching
of the fix in at Square. Now
`./build-support/jenkins/build.sh` runs green as well.

Reviewed at https://reviews.apache.org/r/40334/