Replace manual Forwarding* with `@Forward`.
This is a bit of a straw man.  While working on
https://issues.apache.org/jira/browse/AURORA-987 I found the need for
`javapoet` and this project spun off the side.

See the `@Forward` README here for more info: https://github.com/perkuno/forward

Although I think the end result is desirable, this change does add a
dependency on a personal project.  I fancied up the presentation with a
custom org, but its still a personal project. I'd be happy enough to
move the `@Forward` code over to aurora, but it would require a seperate
gradle module to ensure the annotation processor is compiled ahead of
the main module that would use it.

So kick the tires and let me know what you think.

 build.gradle                                                                  |   6 +-
 src/main/java/org/apache/aurora/scheduler/storage/ForwardingStore.java        | 185 ------------------------------
 src/main/java/org/apache/aurora/scheduler/storage/log/WriteAheadStorage.java  |  13 ++-
 src/test/java/org/apache/aurora/scheduler/thrift/aop/ForwardingThrift.java    | 309 --------------------------------------------------
 src/test/java/org/apache/aurora/scheduler/thrift/aop/MockDecoratedThrift.java |   5 +-
 5 files changed, 20 insertions(+), 498 deletions(-)

Testing Done:
Green locally `./build-support/jenkins/build.sh`.

An example of the generated code:
```java
@Generated("uno.perk.forward.apt.Processor")
class MockDecoratedThriftForwarder implements AnnotatedAuroraAdmin {
  protected final AnnotatedAuroraAdmin annotatedAuroraAdmin;

  MockDecoratedThriftForwarder(AnnotatedAuroraAdmin annotatedAuroraAdmin) {
    this.annotatedAuroraAdmin = Objects.requireNonNull(annotatedAuroraAdmin);
  }

  @Override
  public Response getRoleSummary() throws TException {
    return this.annotatedAuroraAdmin.getRoleSummary();
  }
...
```

Reviewed at https://reviews.apache.org/r/40786/