Update leader redirection logic to return an error page if there is no leading scheduler.
We'd like to serve Aurora through DNS rather than against the singleton serverset so that when
there is no leader users get a friendly error page rather than a generic error (which inevitably
leads to them coming to ask for support). In order to accomplish this, I've updated the
`LeaderRedirectFilter` to return a 503 with an html body in the event there is no elected leader.

Reviewed at https://reviews.apache.org/r/41534/