Fixup ZooKeeperTestServer restartNetwork.
With the upgrade to 3.4.2 and the switch from `NIOServerCnxn.Factory` to
`NIOServerCnxnFactory` came a change in the server internals.  Calling
`connectionFactory.startup(zooKeeperServer)` now eventually calls
`ZooKeeperServer.startup` which calls `SessionTrackerImpl.start`.  This
last is a problem since `SessionTrackerImpl` is a `Thread` and so throws
`IllegalThreadStateException`.

Reviewed at https://reviews.apache.org/r/41894/