Kill flaky TaskObserverTest.
Previously, a mock threading.Event was waited on in one thread
and the count of waits was read in another thread.  Most thread
memory models do not guaranty reads are fresh in this scenario
unless there is a memory barrier of some sort forcing per-cpu
caches to be flushed.

Since the test really only verified correct conversion of a poll
interval to fractional seconds - kill the test as not pulling its
weight.

Bugs closed: AURORA-1570

Reviewed at https://reviews.apache.org/r/41915/