Fix flaky `ServerSetImplTest` test.
The `testUnwatchOnException` test method uses a forced
InterruptedException to test that watches are un-registered.  In doing
so, the test inadvertantly set the interrupt bit for the test runner
thread, poisoning subsequent tests that invoked blocking code.  The
poisoning was only evident when the test methods were not run in lexical
order, which is the case in the vagrant vm.  This fix explicitly clears
the interrupt bit for the test thread with an explanation of why this is
done.

Testing Done:
Before the fix, this consistent error in the vagrant VM:
```
vagrant@aurora:~/aurora$ ./gradlew --rerun-tasks commons:test --tests org.apache.aurora.common.zookeeper.ServerSetImplTest
...
org.apache.aurora.common.zookeeper.ServerSetImplTest > testOrdering FAILED
    org.apache.aurora.common.net.pool.DynamicHostSet$MonitorException at ServerSetImplTest.java:155
        Caused by: org.apache.aurora.common.zookeeper.Group$WatchException at ServerSetImplTest.java:155
            Caused by: org.apache.aurora.common.zookeeper.Group$JoinException at ServerSetImplTest.java:155
                Caused by: java.lang.InterruptedException at ServerSetImplTest.java:155
```

Green after the fix in the vm and when run normally on my machine.

Bugs closed: AURORA-1574

Reviewed at https://reviews.apache.org/r/42102/