Add `--show-error` to curl when bootstrapping thrift.
From the curl documentation:
````
-S, --show-error

When used with -s it makes curl show an error message if it fails.
````

It's possible for curl to fail when grabbing the tarball or patch and this will
show users why it failed.

Testing Done:
Ran `make` in the `build-support/thrift` directory.

Reviewed at https://reviews.apache.org/r/42225/