Fixup broken jmh benchmarks.
The `TierConfig` binding fix for the `SchedulingBenchmarks` in
https://reviews.apache.org/r/42073 was also needed for the
`StatusUpdateBenchmarks` and a binding for `ConfigurationManager` was
missing for the `ThriftApiBenchmarks` as-of
https://reviews.apache.org/r/41711/.

While debugging these failures I noticed a new version of jmh had been
released recently, so also upgraded to that.  No changes of import, but
the changelog can be read here:
  http://hg.openjdk.java.net/code-tools/jmh/

Testing Done:
Both of these now work on master:
`./gradlew jmh -Pbenchmarks='StatusUpdateBenchmark.*'`
`./gradlew jmh -Pbenchmarks='ThriftApiBenchmark.*'`

I also checked that a full run of `./gradlew jmh` was now green for
all benchmarks.

Reviewed at https://reviews.apache.org/r/42475/