Enable H2 logging to slf4j.
On a test cluster with DbTaskStore enabled there are several lines in the log
that look like:
````
2016-01-26 13:07:14 jdbc[15]: exception
````
There is no other information with these lines. This is a result of setting
`TRACE_LEVEL_SYSTEM_OUT` to `1` for H2. This will print out the error message
but not the associated throwable:
https://github.com/h2database/h2database/blob/a932268843ac84c7a665e427167ff2eb291d6b8e/h2/src/main/org/h2/message/TraceSystem.java#L228

The SLF4J implementation of tracing in H2 does not suffer from this restriction.

Reviewed at https://reviews.apache.org/r/42845/