Increase throughput of DbTaskStore
Profiling master indicated that the bottleneck was MyBatis populating ResultSets
and populating the resulting objects. This patch removes subselects, which
reduces the number of ResultSets and removes the population of an object via a
constructor which is slower than populating an object via setters.

Reviewed at https://reviews.apache.org/r/43457/