Fixup install docs.
This set of fixes eliminates the experimental rpm warnings and updates
the rpm instructions to use the officially released packages.  The deb
instructions are updated as well, in particular to take advantage of the
mesosphere deb repository and with movement of special dep installations
to the appropriate sections requiring them.

This fix RB does not address the new Debian Jessie debs, instead
focusing on getting the existing instructions corrected.

Bugs closed: AURORA-1647

Reviewed at https://reviews.apache.org/r/45167/