Simplify `Credentials`; kill `ZooKeeperClient` dep.
The Curator discovery code will need to be configured from the same
command line flags and code as the commons discovery code.  This
simplifies Credentials to be a simple struct and adapts from the
`Credentials.NONE` null-object to use of `Optional` in consumers.

Bugs closed: AURORA-1468

Reviewed at https://reviews.apache.org/r/46133/