Fix bug when fetching a task with multiple assigned ports.
Because we did not include the ids for task ports and task events in the query when mybatis tried
to map the results, there was no way to identify the task events returned in each additional row
for ports as distinct. This resulted in each subsequent task fetch tacking on extra task events.

Bugs closed: AURORA-1672

Reviewed at https://reviews.apache.org/r/46506/