Close `PathChildrenCache` before its framework.
Previously these lifecycles were modeled as independent when, in fact,
a `CuratorFramework`'s clients must be closed befor it is closed to
prevent errors in the clients from attempting to use a closed
`CuratorFramework`.

The proof that closing was always safe already existed in
`CuratorServiceGroupMonitorTest::testExceptionalLifecycle`, but this
safety is now documented and more explicitly tested.

Bugs closed: AURORA-1729

Reviewed at https://reviews.apache.org/r/49578/