Support TBinaryProtocol over HTTP
This replaces the `TServlet` servlet from thrift with our own servlet which
dispatches thrift responses based on the content type of the request. This
enables a client to use either the thrift json protocol or the binary protocol
when communicating with the scheduler.

Without this patch the current behaviour is:
- Regardless of content type of the request, assume the request is json and
  return json with a `application/x-thrift` content type.

With this patch the behaviour becomes:
- A request with no content type header, or a type of `application/x-thrift` or
  `application/json` or `application/vnd.apache.thrift.json` is assumed to be
  JSON.
- A request with a content type header of `application/vnd.apache.thrift.binary`
  is assumed to be binary.
- A request with an `Accept` header of `application/vnd.apache.thrift.binary`
  will have a binary response.
- A request with any other `Accept` header will have a JSON response.

Bugs closed: AURORA-1743

Reviewed at https://reviews.apache.org/r/50685/