Fix thermos killing heuristic to permit setuid(2).
Previously this process killing heuristic would not allow killing of a process
if the uid it was launched with differs from the real uid of the currently
running process. The logic is too conservative because it doesn't factor in
that a process launched as root can use `setuid(2)` to change it's real uid.

This patch fixes the heuristic by permitting killing of a process launched as
root but the real uid is now not root.

Bugs closed: AURORA-1753

Reviewed at https://reviews.apache.org/r/51348/