Refactor of Webhook and no longer posting entire task state via webhook on scheduler restart
This is a refactor with addition of HttpClient injected into Webhook class as opposed to previous usage of lower level HtttpURLConnection objects. Additionally due to peformance issues, it is unncessary to POST entire task state to webhook endpoint on every scheduler restart so that is removed in this patch.

Bugs closed: AURORA-1772

Reviewed at https://reviews.apache.org/r/51980/