Upgrade to latest pants 1.2.0 pre-release.
This pulls in a fix that went in in 1.2.0dev7 that expands the fetch
timeout for the thrift binary used for our python builds from 1 second
to 30. This should provide relief for a top cause of build flakiness.

Release notes are here: http://pantsbuild.github.io/notes-master.html

The latest has dependencies that use cffi which in turn depends on
libffi and so this is added to the Vagrant image.

Additionally, the mesos slave `--executor_registration_timeout` is set
to 1 minute as a temporary workaround to alleviate flaky e2e test
failures related to: https://issues.apache.org/jira/browse/MESOS-6004.

Bugs closed: AURORA-1784

Reviewed at https://reviews.apache.org/r/52367/