Reduce offer_filter_duration to 0secs in Vagrant
This was more a random experiment on my end. We use this value in production as
our clusters are rather small and it can happen that `#instances in a job >> #agents`,
so a smaller filter duration increases scheduling performance. Turns out this is also
slightly helpful in our e2e tests:

Before the patch:

    real   	27m52.457s
    user   	0m2.308s
    sys    	0m0.938s

With the patch:

    real   	23m56.899s
    user   	0m2.277s
    sys    	0m0.884s

Both experiments were preceeded with a `vagrant destroy && vagrant up`.

The improvement is minimal and could probably also due to other external factors as I
have made only one run each.

Reviewed at https://reviews.apache.org/r/52581/