Enable per task volume mounts via scheduler API
This allows users to specify volume mounts for tasks using the unified
containerizer if the operator permits them. This is analogous to enabling docker
parameters per task and using the `--volume` parameter.

This does not include the needed DSL changes or an e2e test which will be in a
subsequent diff.

Bugs closed: AURORA-1107

Reviewed at https://reviews.apache.org/r/52588/