Change job updates to rely on `health-checks` rather than on `watch_secs`.
Make RUNNING a first class state to indicate that the task is running
and is healthy. It is achieved by introducing a new configuration
parameter `min_consecutive_successes`, which will dictate when to move
a task into RUNNING state.

With this change, it is possible to set the `watch_secs` to 0, so that
updates are purely based on the task's health, rather than relying on
watching the task to in RUNNING state for a pre-determined timeout.

Testing Done:
buils-support/jenkins/build.sh
sh ./src/test/sh/org/apache/aurora/e2e/test_end_to_end.sh

Bugs closed: AURORA-1225

Reviewed at https://reviews.apache.org/r/53590/