Adopt built-in string formatting in Preconditions.checkState and our logger.
Inspired by https://reviews.apache.org/r/53928/ this replaces many usages of
`String.format` with the built-in formatting in `Preconditions.checkState` and
our logger. This has the advantage that the formatting is only done when
necessary. A couple of other usages are replaced with `String.join` or simple
string concatenation which tends to be faster than the more powerful
`Sting.format`.

Reviewed at https://reviews.apache.org/r/53933/