SlaUtil::percentile percentiles interpolation
Modification to SlaUtil::percentile to compute percentiles by interpolation

The calculation of mttX (median-time-to-X) depends on the computation of percentile values. The
current implementation does not behave nicely with a small sample size. For instance, for a given
sample set of  {50, 150}, 50-percentile is reported to be 50. Although, 100 seems a more appropriate
return value.

This patch uses Guava's Quantile for interpolation.

Reviewed at https://reviews.apache.org/r/54106/