Get pants using the same thrift binary as gradle.
This required an upgrade to the latest pants dev release to correct
an issue with the setup.py binary packager we use to generate sdists.

This is for sanity sake, and, once the TODO in
`build-support/thrift/prepare_binary.sh` is addressed, it will also
allow pants-patch-free addition of new platforms (thinking ARM).

Reviewed at https://reviews.apache.org/r/53836/