Log process sampling failures with debug severity
The observer's logs consist of lots of warnings about being unable to find PIDs.
This is expected when running with the PID isolator, or when checkpoints are out
of date (e.g. after processes were killed by the OOM).

    W0116 14:42:54.694221 3253 process_collector_psutil.py:75] Error during process sampling: psutil.NoSuchProcess process no longer exists (pid=27727)
    W0116 14:42:54.717905 3253 process_collector_psutil.py:42] Error during process sampling [pid=10960]: psutil.NoSuchProcess process no longer exists (pid=10960)
    W0116 14:42:54.718089 3253 process_collector_psutil.py:75] Error during process sampling: psutil.NoSuchProcess process no longer exists (pid=10960)
    W0116 14:42:54.718245 3253 process_collector_psutil.py:42] Error during process sampling [pid=10026]: psutil.NoSuchProcess process no longer exists (pid=10026)
    W0116 14:42:54.718334 3253 process_collector_psutil.py:75] Error during process sampling: psutil.NoSuchProcess process no longer exists (pid=10026)

This change adopts the proposal of David Robinson to decrease the severity level
to debug.

Bugs closed: AURORA-1541

Reviewed at https://reviews.apache.org/r/55578/