Improve `thriftw` robustness.
Now the selected thrift is checked both for the proper version and
support of the gen langs Aurora requires. In addition, all thrifts on
the `PATH` are and an existing locally built thrift is always verified
to protect Aurora thrift requirement changes (if we ever add a gen lang
again).

Bugs closed: AURORA-1875

Reviewed at https://reviews.apache.org/r/55536/